Update requests to interact with NVD API2.0
NIST deprecated their API1.0 in December 2023 and replaced it with
API2.0. This change breaks line 144 of the get_cvss_score() method in
vuln_analyzer.py. Also had to increase the wait time between requests
to 5 seconds since it seems that the rate limiting is harsher with the
new API.

Changes to the API can be found here:
https://nvd.nist.gov/general/news/api-20-announcements
https://nvd.nist.gov/developers/vulnerabilities