styles/syntax: Major syntax and style changes to passive_attacks directory.
Changes to All Passive Attack Files:
Added google style python doc strings to each method.
Added type hinting for most (if not all) variables and function return types.
Changed how some functions parameters are listed to lessen line length.

VulnAnalyzer:
Changed the version check for Conpot within the `get_release_date()` method to `version_tag > "0.2.2"` since  that's when Conpot stopped using just the release number.

StaticHoney:
Changed `cve_ids` to be a `set` instead of a list as we have to check whether we've added it anyway so the O(1) lookup of the ID's is beneficial.
Created the `FilteredRes` type alias for interacting with bandit's result JSON output.
Created the `FilteredJSON` type alias for logging the `FilterRes` to a JSON log file.