attack: Revamped DoS attack.
Rewrote the attacks in Go as Python's current threading implementation with the GIL isn't the best solution for something that needs to output data as much as possible.
Combined dos_all_open_ports and dos as they do the same thing but now both will run nmap to get all the ports needed.
Updated README.md to add `sudo` to run the tool.