detection: Implement the honeypot detection into codebase.
Using the code from [this](https://github.com/honeynet/honeyscanner/pull/38) PR, I've implemented it to work within the current codebase.
Honeyscanner runs with the detection being a default so we don't need to know the type, version, or open ports to pass in as arguments to the program as it will get those automatically.